Adding v1 endpoint calls to hnslib.
