+++ CODEOWNERS
@@ -4,8 +4,12 @@ approvers:
   - sbangari@microsoft.com
   - souravdas@microsoft.com
   - ppereira@microsoft.com
+  - debjitmondal@microsoft.com
+  - daschott@microsoft.com
 
 reviewers:
   - sbangari@microsoft.com
   - souravdas@microsoft.com
   - ppereira@microsoft.com
+  - debjitmondal@microsoft.com
+  - daschott@microsoft.com
